+++ CHANGES.rst
@@ -17,6 +17,9 @@ Unreleased
   fields :attr:`~UpdatedFeed.new` and :attr:`~UpdatedFeed.modified`
   became optional.
   (:issue:`96`)
+* Fix bug in :mod:`~reader.plugins.entry_dedupe` that failed updates
+  of feeds with multiple *new* duplicates of the same issue.
+  (:issue:`292`)
 * In the CLI, don't suppress the traceback of :exc:`ReaderError`,
   since it would also suppress it for bugs.
 * In the CLI, stop using deprecated :func:`click.get_terminal_size`.